requirements.txt
opencv-python
numpy
pyserial
openai
ultralytics
python-dotenv
pyyaml
